+++ src/main/java/net/id/paradiselost/client/rendering/particle/CherineFlameParticle.java
@@ -12,27 +12,27 @@ public class CherineFlameParticle extends AbstractSlowingParticle {
     CherineFlameParticle(ClientWorld clientWorld, double d, double e, double f, double g, double h, double i) {
         super(clientWorld, d, e, f, g, h, i);
     }
-    
+
     @Override
     public ParticleTextureSheet getType() {
         return ParticleTextureSheet.PARTICLE_SHEET_OPAQUE;
     }
-    
+
     @Override
     public void move(double dx, double dy, double dz) {
         this.setBoundingBox(this.getBoundingBox().offset(dx, dy, dz));
         this.repositionFromBoundingBox();
     }
-    
+
     @Override
     public float getSize(float tickDelta) {
-        float f = ((float) this.age + tickDelta) / (float) this.maxAge;
+        float f = (this.age + tickDelta) / this.maxAge;
         return this.scale * (1.0F - f * f * 0.5F);
     }
-    
+
     @Override
     public int getBrightness(float tint) {
-        float f = ((float) this.age + tint) / (float) this.maxAge;
+        float f = (this.age + tint) / this.maxAge;
         f = MathHelper.clamp(f, 0.0F, 1.0F);
         int i = super.getBrightness(tint);
         int j = i & 255;
@@ -41,18 +41,18 @@ public class CherineFlameParticle extends AbstractSlowingParticle {
         if (j > 240) {
             j = 240;
         }
-        
+
         return j | k << 16;
     }
-    
+
     @Environment(EnvType.CLIENT)
     public static class DefaultFactory implements ParticleFactory<SimpleParticleType> {
         private final SpriteProvider spriteProvider;
-        
+
         public DefaultFactory(SpriteProvider spriteProvider) {
             this.spriteProvider = spriteProvider;
         }
-        
+
         @Override
         public Particle createParticle(SimpleParticleType defaultParticleType, ClientWorld clientWorld, double d, double e, double f, double g, double h, double i) {
             CherineFlameParticle flameParticle = new CherineFlameParticle(clientWorld, d, e, f, g, h, i);
